@@ -18,16 +18,18 @@ SASS_GEMSPEC = Gem::Specification.new do |spec|
       command line tool or a web-framework plugin.
     END
 
-  spec.required_ruby_version = '>= 1.8.7'
+  spec.required_ruby_version = '>= 2.0.0'
+
+  spec.add_runtime_dependency 'sass-listen', '~> 3.0.7'
+  
   spec.add_development_dependency 'yard', '~> 0.8.7.6'
   spec.add_development_dependency 'redcarpet', '~> 2.3.0'
   spec.add_development_dependency 'minitest', '>= 5'
 
   readmes = Dir['*'].reject{ |x| x =~ /(^|[^.a-z])[a-z]+/ || x == "TODO" }
   spec.executables = ['sass', 'sass-convert', 'scss']
-  spec.files = Dir['rails/init.rb', 'lib/**/*', 'vendor/**/*',
-    'bin/*', 'test/**/*', 'extra/**/*', 'Rakefile', 'init.rb',
-    '.yardopts'] + readmes
+  spec.files = Dir['rails/init.rb', 'lib/**/*', 'bin/*', 'test/**/*',
+    'extra/**/*', 'Rakefile', 'init.rb', '.yardopts'] + readmes
   spec.homepage = 'http://sass-lang.com/'
   spec.has_rdoc = false
   spec.test_files = Dir['test/**/*_test.rb']
