@@ -1,5 +1,4 @@
 require 'sass/script/lexer'
-require 'sass/script/css_variable_warning'
 
 module Sass
   module Script
@@ -35,7 +34,6 @@ module Sass
         @allow_extra_text = options.delete(:allow_extra_text)
         @lexer = lexer_class.new(str, line, offset, options)
         @stop_at = nil
-        @css_variable_warning = nil
       end
 
       # Parses a SassScript expression within an interpolated segment (`#{}`).
@@ -52,9 +50,7 @@ module Sass
         start_pos = Sass::Source::Position.new(line, offset - 2)
         expr = assert_expr :expr
         assert_tok :end_interpolation
-        expr = Sass::Script::Tree::Interpolation.new(
-          nil, expr, nil, false, false, :warn_for_color => warn_for_color)
-        check_for_interpolation expr
+        expr = Sass::Script::Tree::Interpolation.new(expr, warn_for_color)
         expr.options = @options
         node(expr, start_pos)
       rescue Sass::SyntaxError => e
@@ -64,23 +60,12 @@ module Sass
 
       # Parses a SassScript expression.
       #
-      # @param css_variable [Boolean] Whether this is the value of a CSS variable.
       # @return [Script::Tree::Node] The root node of the parse tree
       # @raise [Sass::SyntaxError] if the expression isn't valid SassScript
-      def parse(css_variable = false)
-        if css_variable
-          @css_variable_warning = CssVariableWarning.new
-        end
-
+      def parse
         expr = assert_expr :expr
         assert_done
         expr.options = @options
-        check_for_interpolation expr
-
-        if css_variable
-          @css_variable_warning.value = expr
-        end
-
         expr
       rescue Sass::SyntaxError => e
         e.modify_backtrace :line => @lexer.line, :filename => @options[:filename]
@@ -98,7 +83,6 @@ module Sass
         expr = assert_expr :expr
         assert_done
         expr.options = @options
-        check_for_interpolation expr
         expr
       rescue Sass::SyntaxError => e
         e.modify_backtrace :line => @lexer.line, :filename => @options[:filename]
@@ -122,26 +106,10 @@ module Sass
         end
         assert_done
 
-        args.each do |a|
-          check_for_interpolation a
-          a.options = @options
-        end
-
-        keywords.each do |_k, v|
-          check_for_interpolation v
-          v.options = @options
-        end
-
-        if splat
-          check_for_interpolation splat
-          splat.options = @options
-        end
-
-        if kwarg_splat
-          check_for_interpolation kwarg_splat
-          kwarg_splat.options = @options
-        end
-
+        args.each {|a| a.options = @options}
+        keywords.each {|_k, v| v.options = @options}
+        splat.options = @options if splat
+        kwarg_splat.options = @options if kwarg_splat
         return args, keywords, splat, kwarg_splat
       rescue Sass::SyntaxError => e
         e.modify_backtrace :line => @lexer.line, :filename => @options[:filename]
@@ -158,20 +126,10 @@ module Sass
         assert_done
 
         args.each do |k, v|
-          check_for_interpolation k
           k.options = @options
-
-          if v
-            check_for_interpolation v
-            v.options = @options
-          end
+          v.options = @options if v
         end
-
-        if splat
-          check_for_interpolation splat
-          splat.options = @options
-        end
-
+        splat.options = @options if splat
         return args, splat
       rescue Sass::SyntaxError => e
         e.modify_backtrace :line => @lexer.line, :filename => @options[:filename]
@@ -188,20 +146,10 @@ module Sass
         assert_done
 
         args.each do |k, v|
-          check_for_interpolation k
           k.options = @options
-
-          if v
-            check_for_interpolation v
-            v.options = @options
-          end
+          v.options = @options if v
         end
-
-        if splat
-          check_for_interpolation splat
-          splat.options = @options
-        end
-
+        splat.options = @options if splat
         return args, splat
       rescue Sass::SyntaxError => e
         e.modify_backtrace :line => @lexer.line, :filename => @options[:filename]
@@ -221,7 +169,6 @@ module Sass
         end
 
         expr = assert_expr :funcall
-        check_for_interpolation expr
         expr.options = @options
         @lexer.unpeek!
         expr
@@ -232,12 +179,12 @@ module Sass
 
       # Parses a SassScript expression.
       #
+      # @overload parse(str, line, offset, filename = nil)
       # @return [Script::Tree::Node] The root node of the parse tree
       # @see Parser#initialize
       # @see Parser#parse
-      def self.parse(value, line, offset, options = {})
-        css_variable = options.delete :css_variable
-        new(value, line, offset, options).parse(css_variable)
+      def self.parse(*args)
+        new(*args).parse
       end
 
       PRECEDENCE = [
@@ -249,8 +196,6 @@ module Sass
       ]
 
       ASSOCIATIVE = [:plus, :times]
-
-      VALID_CSS_OPS = [:comma, :single_eq, :space, :div]
 
       class << self
         # Returns an integer representing the precedence
@@ -281,20 +226,8 @@ module Sass
         def production(name, sub, *ops)
           class_eval <<RUBY, __FILE__, __LINE__ + 1
             def #{name}
-              interp = try_ops_after_interp(#{ops.inspect}, #{name.inspect})
-              return interp if interp
               return unless e = #{sub}
               while tok = try_toks(#{ops.map {|o| o.inspect}.join(', ')})
-                if interp = try_op_before_interp(tok, e)
-                  other_interp = try_ops_after_interp(#{ops.inspect}, #{name.inspect}, interp)
-                  return interp unless other_interp
-                  return other_interp
-                end
-
-                if @css_variable_warning && !VALID_CSS_OPS.include?(tok.type)
-                  @css_variable_warning.warn!
-                end
-
                 e = node(Tree::Operation.new(e, assert_expr(#{sub.inspect}), tok.type),
                          e.source_range.start_pos)
               end
@@ -306,12 +239,8 @@ RUBY
         def unary(op, sub)
           class_eval <<RUBY, __FILE__, __LINE__ + 1
             def unary_#{op}
-              return #{sub} unless tok = try_tok(:#{op})
-              interp = try_op_before_interp(tok)
-              return interp if interp
+              return #{sub} unless try_tok(:#{op})
               start_pos = source_position
-
-              @css_variable_warning.warn! if @css_variable_warning
               node(Tree::UnaryOperation.new(assert_expr(:unary_#{op}), :#{op}), start_pos)
             end
 RUBY
@@ -333,12 +262,11 @@ RUBY
 
       def map
         start_pos = source_position
-        e = interpolation
+        e = space
         return unless e
         return list e, start_pos unless @lexer.peek && @lexer.peek.type == :colon
 
         pair = map_pair(e)
-        @css_variable_warning.warn! if @css_variable_warning
         map = node(Sass::Script::Tree::MapLiteral.new([pair]), start_pos)
         while try_tok(:comma)
           pair = map_pair
@@ -349,14 +277,14 @@ RUBY
       end
 
       def map_pair(key = nil)
-        return unless key ||= interpolation
+        return unless key ||= space
         assert_tok :colon
-        return key, assert_expr(:interpolation)
+        return key, assert_expr(:space)
       end
 
       def expr
         start_pos = source_position
-        e = interpolation
+        e = space
         return unless e
         list e, start_pos
       end
@@ -364,118 +292,16 @@ RUBY
       def list(first, start_pos)
         return first unless @lexer.peek && @lexer.peek.type == :comma
 
-        list = node(Sass::Script::Tree::ListLiteral.new([first], :comma), start_pos)
-        while (tok = try_tok(:comma))
-          element_before_interp = list.elements.length == 1 ? list.elements.first : list
-          if (interp = try_op_before_interp(tok, element_before_interp))
-            other_interp = try_ops_after_interp([:comma], :expr, interp)
-            return interp unless other_interp
-            return other_interp
-          end
-          return list unless (e = interpolation)
+        list = node(Sass::Script::Tree::ListLiteral.new([first], separator: :comma), start_pos)
+        while try_tok(:comma)
+          return list unless (e = space)
           list.elements << e
           list.source_range.end_pos = list.elements.last.source_range.end_pos
         end
         list
       end
 
-      production :equals, :interpolation, :single_eq
-
-      def try_op_before_interp(op, prev = nil, after_interp = false)
-        return unless @lexer.peek && @lexer.peek.type == :begin_interpolation
-        unary = !prev && !after_interp
-        wb = @lexer.whitespace?(op)
-        str = literal_node(Script::Value::String.new(Lexer::OPERATORS_REVERSE[op.type]),
-                           op.source_range)
-
-        deprecation =
-          case op.type
-          when :comma; :potential
-          when :div, :single_eq; :none
-          when :plus; unary ? :none : :immediate
-          when :minus; @lexer.whitespace?(@lexer.peek) ? :immediate : :none
-          else; :immediate
-          end
-
-        interp = node(
-          Script::Tree::Interpolation.new(
-            prev, str, nil, wb, false, :originally_text => true, :deprecation => deprecation),
-          (prev || str).source_range.start_pos)
-        interpolation(interp)
-      end
-
-      def try_ops_after_interp(ops, name, prev = nil)
-        return unless @lexer.after_interpolation?
-        op = try_toks(*ops)
-        return unless op
-        interp = try_op_before_interp(op, prev, :after_interp)
-        return interp if interp
-
-        wa = @lexer.whitespace?
-        str = literal_node(Script::Value::String.new(Lexer::OPERATORS_REVERSE[op.type]),
-                           op.source_range)
-        str.line = @lexer.line
-
-        deprecation =
-          case op.type
-          when :comma; :potential
-          when :div, :single_eq; :none
-          when :minus; @lexer.whitespace?(op) ? :immediate : :none
-          else; :immediate
-          end
-        interp = node(
-          Script::Tree::Interpolation.new(
-            prev, str, assert_expr(name), false, wa,
-            :originally_text => true, :deprecation => deprecation),
-          (prev || str).source_range.start_pos)
-        interp
-      end
-
-      def interpolation(first = space)
-        e = first
-        while (interp = try_tok(:begin_interpolation))
-          wb = @lexer.whitespace?(interp)
-          char_before = @lexer.char(interp.pos - 1)
-          mid = without_css_variable_warning {assert_expr :expr}
-          assert_tok :end_interpolation
-          wa = @lexer.whitespace?
-          char_after = @lexer.char
-
-          after = space
-          before_deprecation = e.is_a?(Script::Tree::Interpolation) ? e.deprecation : :none
-          after_deprecation = after.is_a?(Script::Tree::Interpolation) ? after.deprecation : :none
-
-          deprecation =
-            if before_deprecation == :immediate || after_deprecation == :immediate ||
-               # Warn for #{foo}$var and #{foo}(1) but not #{$foo}1.
-               (after && !wa && char_after =~ /[$(]/) ||
-               # Warn for $var#{foo} and (a)#{foo} but not a#{foo}.
-               (e && !wb && is_unsafe_before?(e, char_before))
-              :immediate
-            else
-              :potential
-            end
-
-          e = node(
-            Script::Tree::Interpolation.new(e, mid, after, wb, wa, :deprecation => deprecation),
-            (e || interp).source_range.start_pos)
-        end
-        e
-      end
-
-      # Returns whether `expr` is unsafe to include before an interpolation.
-      #
-      # @param expr [Node] The expression to check.
-      # @param char_before [String] The character immediately before the
-      #   interpolation being checked (and presumably the last character of
-      #   `expr`).
-      # @return [Boolean]
-      def is_unsafe_before?(expr, char_before)
-        return char_before == ')' if is_safe_value?(expr)
-
-        # Otherwise, it's only safe if it was another interpolation.
-        !expr.is_a?(Script::Tree::Interpolation)
-      end
+      production :equals, :space, :single_eq
 
       # Returns whether `expr` is safe as the value immediately before an
       # interpolation.
@@ -500,7 +326,7 @@ RUBY
         if arr.size == 1
           arr.first
         else
-          node(Sass::Script::Tree::ListLiteral.new(arr, :space), start_pos)
+          node(Sass::Script::Tree::ListLiteral.new(arr, separator: :space), start_pos)
         end
       end
 
@@ -517,20 +343,57 @@ RUBY
       unary :not, :ident
 
       def ident
-        return funcall unless @lexer.peek && @lexer.peek.type == :ident
-        return if @stop_at && @stop_at.include?(@lexer.peek.value)
+        return funcall unless (first = @lexer.peek)
 
-        name = @lexer.next
-        if (color = Sass::Script::Value::Color::COLOR_NAMES[name.value.downcase])
-          literal_node(Sass::Script::Value::Color.new(color, name.value), name.source_range)
-        elsif name.value == "true"
-          literal_node(Sass::Script::Value::Bool.new(true), name.source_range)
-        elsif name.value == "false"
-          literal_node(Sass::Script::Value::Bool.new(false), name.source_range)
-        elsif name.value == "null"
-          literal_node(Sass::Script::Value::Null.new, name.source_range)
+        contents = []
+        if first.type == :ident
+          return if @stop_at && @stop_at.include?(first.value)
+          contents << @lexer.next.value
+        elsif first.type == :begin_interpolation
+          @lexer.next # Move through :begin_interpolation
+          contents << assert_expr(:expr)
+          assert_tok(:end_interpolation)
         else
-          literal_node(Sass::Script::Value::String.new(name.value, :identifier), name.source_range)
+          return funcall
+        end
+
+        while (tok = @lexer.peek)
+          break if @lexer.whitespace_before?(tok)
+
+          if tok.type == :ident
+            contents << @lexer.next.value
+            next
+          elsif tok.type == :funcall
+            contents << @lexer.next.value
+            args, keywords, splat, kwarg_splat = fn_arglist
+            assert_tok(:rparen)
+            return node(Script::Tree::Funcall.new(contents, args, keywords, splat, kwarg_splat),
+                 tok.source_range.start_pos, source_position)
+          end
+
+          break unless try_tok(:begin_interpolation)
+          contents << assert_expr(:expr)
+          assert_tok(:end_interpolation)
+        end
+
+        if contents.length > 1 || contents.first.is_a?(Sass::Script::Tree::Node)
+          return node(
+            Sass::Script::Tree::StringInterpolation.new(contents, :identifier),
+            first.source_range.start_pos)
+        end
+
+        if (color = Sass::Script::Value::Color::COLOR_NAMES[first.value.downcase])
+          literal_node(Sass::Script::Value::Color.new(color, first.value), first.source_range)
+        elsif first.value == "true"
+          literal_node(Sass::Script::Value::Bool.new(true), first.source_range)
+        elsif first.value == "false"
+          literal_node(Sass::Script::Value::Bool.new(false), first.source_range)
+        elsif first.value == "null"
+          literal_node(Sass::Script::Value::Null.new, first.source_range)
+        else
+          literal_node(
+            Sass::Script::Value::String.new(first.value, :identifier),
+            first.source_range)
         end
       end
 
@@ -539,7 +402,7 @@ RUBY
         return raw unless tok
         args, keywords, splat, kwarg_splat = fn_arglist
         assert_tok(:rparen)
-        node(Script::Tree::Funcall.new(tok.value, args, keywords, splat, kwarg_splat),
+        node(Script::Tree::Funcall.new([tok.value], args, keywords, splat, kwarg_splat),
           tok.source_range.start_pos, source_position)
       end
 
@@ -549,12 +412,12 @@ RUBY
         else
           return [], nil unless try_tok(:lparen)
         end
-        return [], nil if try_tok(:rparen)
 
         res = []
         splat = nil
         must_have_default = false
         loop do
+          break if peek_tok(:rparen)
           c = assert_tok(:const)
           var = node(Script::Tree::Variable.new(c.value), c.source_range)
           if try_tok(:colon)
@@ -579,18 +442,14 @@ RUBY
       end
 
       def mixin_arglist
-        arglist(:interpolation, "mixin argument")
+        arglist(:space, "mixin argument")
       end
 
       def arglist(subexpr, description)
         args = []
         keywords = Sass::Util::NormalizedMap.new
-        e = send(subexpr)
-
-        return [args, keywords] unless e
-
         splat = nil
-        loop do
+        while (e = send(subexpr))
           if @lexer.peek && @lexer.peek.type == :colon
             name = e
             @lexer.expected!("comma") unless name.is_a?(Tree::Variable)
@@ -611,13 +470,12 @@ RUBY
             elsif !keywords.empty?
               raise SyntaxError.new("Positional arguments must come before keyword arguments.")
             end
-
             args << e if e
           end
 
           return args, keywords, splat unless try_tok(:comma)
-          e = assert_expr(subexpr, description)
         end
+        return args, keywords
       end
 
       def raw
@@ -628,15 +486,66 @@ RUBY
 
       def special_fun
         first = try_tok(:special_fun)
-        return paren unless first
-        str = literal_node(first.value, first.source_range)
-        return str unless try_tok(:string_interpolation)
-        mid = without_css_variable_warning {assert_expr :expr}
-        assert_tok :end_interpolation
-        last = assert_expr(:special_fun)
+        return square_list unless first
+
+        unless try_tok(:string_interpolation)
+          return literal_node(first.value, first.source_range)
+        end
+
+        contents = [first.value.value]
+        begin
+          contents << assert_expr(:expr)
+          assert_tok :end_interpolation
+          contents << assert_tok(:special_fun).value.value
+        end while try_tok(:string_interpolation)
+
         node(
-          Tree::Interpolation.new(str, mid, last, false, false),
+          Tree::StringInterpolation.new(contents, :identifier),
           first.source_range.start_pos)
+      end
+
+      def square_list
+        start_pos = source_position
+        return paren unless try_tok(:lsquare)
+
+        space_start_pos = source_position
+        e = or_expr
+        separator = nil
+        if e
+          elements = [e]
+          while (e = or_expr)
+            elements << e
+          end
+
+          # If there's a comma after a space-separated list, it's actually a
+          # space-separated list nested in a comma-separated list.
+          if try_tok(:comma)
+            e = if elements.length == 1
+                  elements.first
+                else
+                  node(
+                    Sass::Script::Tree::ListLiteral.new(elements, separator: :space),
+                    space_start_pos)
+                end
+            elements = [e]
+
+            while (e = space)
+              elements << e
+              break unless try_tok(:comma)
+            end
+            separator = :comma
+          else
+            separator = :space if elements.length > 1
+          end
+        else
+          elements = []
+        end
+
+        assert_tok(:rsquare)
+        end_pos = source_position
+
+        node(Sass::Script::Tree::ListLiteral.new(elements, separator: separator, bracketed: true),
+             start_pos, end_pos)
       end
 
       def paren
@@ -646,29 +555,34 @@ RUBY
         e.force_division! if e
         end_pos = source_position
         assert_tok(:rparen)
-
-        @css_variable_warning.warn! if @css_variable_warning
-        e || node(Sass::Script::Tree::ListLiteral.new([], nil), start_pos, end_pos)
+        e || node(Sass::Script::Tree::ListLiteral.new([]), start_pos, end_pos)
       end
 
       def variable
         start_pos = source_position
         c = try_tok(:const)
         return string unless c
-
-        @css_variable_warning.warn! if @css_variable_warning
         node(Tree::Variable.new(*c.value), start_pos)
       end
 
       def string
         first = try_tok(:string)
         return number unless first
-        str = literal_node(first.value, first.source_range)
-        return str unless try_tok(:string_interpolation)
-        mid = without_css_variable_warning {assert_expr :expr}
-        assert_tok :end_interpolation
-        last = assert_expr(:string)
-        node(Tree::StringInterpolation.new(str, mid, last), first.source_range.start_pos)
+
+        unless try_tok(:string_interpolation)
+          return literal_node(first.value, first.source_range)
+        end
+
+        contents = [first.value.value]
+        begin
+          contents << assert_expr(:expr)
+          assert_tok :end_interpolation
+          contents << assert_tok(:string).value.value
+        end while try_tok(:string_interpolation)
+
+        node(
+          Tree::StringInterpolation.new(contents, first.value.type),
+          first.source_range.start_pos)
       end
 
       def number
@@ -683,7 +597,6 @@ RUBY
       def selector
         tok = try_tok(:selector)
         return literal unless tok
-        @css_variable_warning.warn! if @css_variable_warning
         node(tok.value, tok.source_range.start_pos)
       end
 
@@ -700,8 +613,7 @@ RUBY
         :default => "expression (e.g. 1px, bold)",
         :mixin_arglist => "mixin argument",
         :fn_arglist => "function argument",
-        :splat => "...",
-        :special_fun => '")"',
+        :splat => "..."
       }
 
       def assert_expr(name, expected = nil)
@@ -723,10 +635,14 @@ RUBY
         @lexer.expected!(names.map {|tok| Lexer::TOKEN_NAMES[tok] || tok}.join(" or "))
       end
 
-      def try_tok(name)
+      def peek_tok(name)
         # Avoids an array allocation caused by argument globbing in the try_toks method.
         peeked = @lexer.peek
-        peeked && name == peeked.type && @lexer.next
+        peeked && name == peeked.type
+      end
+
+      def try_tok(name)
+        peek_tok(name) && @lexer.next
       end
 
       def try_toks(*names)
@@ -771,61 +687,10 @@ RUBY
             range(source_range_or_start_pos, end_pos)
           end
 
-        node.css_variable_warning = @css_variable_warning
         node.line = source_range.start_pos.line
         node.source_range = source_range
         node.filename = @options[:filename]
         node
-      end
-
-      # Runs the given block without CSS variable warnings enabled.
-      #
-      # CSS warnings don't apply within interpolation, so this is used to
-      # disable them.
-      #
-      # @yield []
-      def without_css_variable_warning
-        old_css_variable_warning = @css_variable_warning
-        @css_variable_warning = nil
-        yield
-      ensure
-        @css_variable_warning = old_css_variable_warning
-      end
-
-      # Checks a script node for any immediately-deprecated interpolations, and
-      # emits warnings for them.
-      #
-      # @param node [Sass::Script::Tree::Node]
-      def check_for_interpolation(node)
-        nodes = [node]
-        until nodes.empty?
-          node = nodes.pop
-          unless node.is_a?(Sass::Script::Tree::Interpolation) &&
-                 node.deprecation == :immediate
-            nodes.concat node.children
-            next
-          end
-
-          interpolation_deprecation(node)
-        end
-      end
-
-      # Emits a deprecation warning for an interpolation node.
-      #
-      # @param node [Sass::Script::Tree::Node]
-      def interpolation_deprecation(interpolation)
-        return if @options[:_convert]
-        location = "on line #{interpolation.line}"
-        location << " of #{interpolation.filename}" if interpolation.filename
-        Sass::Util.sass_warn <<WARNING
-DEPRECATION WARNING #{location}:
-\#{} interpolation near operators will be simplified in a future version of Sass.
-To preserve the current behavior, use quotes:
-
-  #{interpolation.to_quoted_equivalent.to_sass}
-
-You can use the sass-convert command to automatically fix most cases.
-WARNING
       end
     end
   end
