@@ -3,15 +3,9 @@ module Sass::Script::Tree
   #
   # @see Interpolation
   class StringInterpolation < Node
-    # @return [Literal] The string literal before this interpolation.
-    attr_reader :before
-
-    # @return [Node] The SassScript within the interpolation
-    attr_reader :mid
-
-    # @return [StringInterpolation, Literal]
-    #     The string literal or string interpolation before this interpolation.
-    attr_reader :after
+    # @return [[Node, String]] The contents of the interpolation. Strings are
+    #   literal text, nodes are interpolated segments.
+    attr_reader :contents
 
     # Whether this is a CSS string or a CSS identifier. The difference is that
     # strings are written with double-quotes, while identifiers aren't.
@@ -20,62 +14,56 @@ module Sass::Script::Tree
     # functions such as `url()`.
     #
     # @return [Symbol] `:string` or `:identifier`
-    def type
-      @before.value.type
-    end
+    attr_reader :type
 
-    # Returns the quote character that should be used to wrap a Sass
-    # representation of this interpolation.
-    def quote
-      quote_for(self) || '"'
-    end
-
-    # Interpolation in a string is of the form `"before #{mid} after"`,
-    # where `before` and `after` may include more interpolation.
+    # Interpolation in a string.
     #
-    # @param before [StringInterpolation, Literal] See {StringInterpolation#before}
-    # @param mid [Node] See {StringInterpolation#mid}
-    # @param after [Literal] See {StringInterpolation#after}
-    def initialize(before, mid, after)
-      @before = before
-      @mid = mid
-      @after = after
+    # @param contents [[Node, String]] See {StringInterpolation#contents}
+    # @param type [Symbol] See {StringInterpolation#type}
+    def initialize(contents, type = :string)
+      @contents = Sass::Util.merge_adjacent_strings(contents)
+      @type = type
     end
 
     # @return [String] A human-readable s-expression representation of the interpolation
     def inspect
-      "(string_interpolation #{@before.inspect} #{@mid.inspect} #{@after.inspect})"
+      "(string_interpolation #{@type.inspect} #{contents.map {|c| c.inspect}.join(', ')})"
     end
 
     # @see Node#to_sass
     def to_sass(opts = {})
-      quote = type == :string ? opts[:quote] || quote_for(self) || '"' : :none
+      quote = type == :string ? opts[:quote] || detected_quote : :none
       opts = opts.merge(:quote => quote)
 
       res = ""
       res << quote if quote != :none
-      res << _to_sass(before, opts)
-      res << '#{' << @mid.to_sass(opts.merge(:quote => nil)) << '}'
-      res << _to_sass(after, opts)
+      contents.each do |c|
+        if c.is_a?(String)
+          c = Sass::Script::Value::String.quote(c, opts).slice(1...-1) if quote != :none
+          res << c
+        else
+          res << '#{' << c.to_sass(opts.merge(:quote => nil)) << '}'
+        end
+      end
       res << quote if quote != :none
       res
     end
 
-    # Returns the three components of the interpolation, `before`, `mid`, and `after`.
+    # Returns the node contents of this interpolation.
     #
     # @return [Array<Node>]
     # @see #initialize
     # @see Node#children
     def children
-      [@before, @mid, @after].compact
+      contents.select {|c| c.is_a?(Sass::Script::Tree::Node)}
     end
 
     # @see Node#deep_copy
     def deep_copy
       node = dup
-      node.instance_variable_set('@before', @before.deep_copy) if @before
-      node.instance_variable_set('@mid', @mid.deep_copy)
-      node.instance_variable_set('@after', @after.deep_copy) if @after
+      node.instance_variable_set('@contents', contents.map do |c|
+        c.is_a?(String) ? c : c.deep_copy
+      end)
       node
     end
 
@@ -87,39 +75,30 @@ module Sass::Script::Tree
     # @return [Sass::Script::Value::String]
     #   The SassScript string that is the value of the interpolation
     def _perform(environment)
-      res = ""
-      before = @before.perform(environment)
-      res << before.value
-      mid = @mid.perform(environment)
-      res << (mid.is_a?(Sass::Script::Value::String) ? mid.value : mid.to_s(:quote => :none))
-      res << @after.perform(environment).value
-      opts(Sass::Script::Value::String.new(res, before.type))
+      opts(Sass::Script::Value::String.new(contents.map do |c|
+        next c if c.is_a?(String)
+        value = c.perform(environment)
+        value.is_a?(Sass::Script::Value::String) ? value.value : value.to_s(:quote => :none)
+      end.join, type))
     end
 
     private
 
-    def _to_sass(string_or_interp, opts)
-      result = string_or_interp.to_sass(opts)
-      opts[:quote] == :none ? result : result.slice(1...-1)
-    end
+    # Returns the best quote to use for this interpolation.
+    #
+    # @return [String]
+    def detected_quote
+      has_double_quote = false
+      contents.each do |c|
+        next if c.is_a?(Sass::Script::Tree::Node)
 
-    def quote_for(string_or_interp)
-      if string_or_interp.is_a?(Sass::Script::Tree::Literal)
-        return nil if string_or_interp.value.value.empty?
-        return '"' if string_or_interp.value.value.include?("'")
-        return "'" if string_or_interp.value.value.include?('"')
-        return nil
+        # Double-quotes take precedence over single quotes, so if any string contains
+        # a single quote we'll definitely end up using double quotes.
+        return '"' if c.include?("'")
+        has_double_quote ||= c.include?('"')
       end
 
-      # Double-quotes take precedence over single quotes.
-      before_quote = quote_for(string_or_interp.before)
-      return '"' if before_quote == '"'
-      after_quote = quote_for(string_or_interp.after)
-      return '"' if after_quote == '"'
-
-      # Returns "'" if either or both insist on single quotes, and nil
-      # otherwise.
-      before_quote || after_quote
+      has_double_quote ? "'" : '"'
     end
   end
 end
