@@ -135,23 +135,7 @@ MESSAGE
       end
 
       def combinator
-        tok(PLUS) || tok(GREATER) || tok(TILDE) || reference_combinator
-      end
-
-      def reference_combinator
-        return unless tok(%r{/})
-        res = '/'
-        ns, name = expr!(:qualified_name)
-        res << ns << '|' if ns
-        res << name << tok!(%r{/})
-
-        location = " of #{@filename}" if @filename
-        Sass::Util.sass_warn <<MESSAGE
-DEPRECATION WARNING on line #{@line}, column #{@offset}#{location}:
-The reference combinator #{res} is deprecated and will be removed in a future release.
-MESSAGE
-
-        res
+        tok(PLUS) || tok(GREATER) || tok(TILDE)
       end
 
       def simple_selector_sequence
@@ -289,35 +273,12 @@ MESSAGE
           elsif s == ':' && PREFIXED_SELECTOR_PSEUDO_CLASSES.include?(deprefixed)
             arg, sel = prefixed_selector_pseudo
           else
-            arg = expr!(:pseudo_args)
+            arg = expr!(:declaration_value).join
           end
 
           tok!(/\)/)
         end
         Selector::Pseudo.new(s == ':' ? :class : :element, name, arg, sel)
-      end
-
-      def pseudo_args
-        arg = expr!(:pseudo_expr)
-        while tok(/,/)
-          arg << ',' << str {ss}
-          arg.concat expr!(:pseudo_expr)
-        end
-        arg
-      end
-
-      def pseudo_expr
-        res = pseudo_expr_token
-        return unless res
-        res << str {ss}
-        while (e = pseudo_expr_token)
-          res << e << str {ss}
-        end
-        res
-      end
-
-      def pseudo_expr_token
-        tok(PLUS) || tok(/[-*]/) || tok(NUMBER) || tok(STRING) || tok(IDENT)
       end
 
       def prefixed_selector_pseudo
